Revision(en) 8196-8197 von branches/8.0-patches zusammengeführt: Another fix for modern MariaDB Connector C versions. ........ By MShark67: @MShark67 Oracle loblocator changes. Fixes to UTF8 lob reading and writing. Threefold performance improvement for "None" and "OnAccess" cachelobmodes.
Note that I've fixed the TZOracleLobStream64.ReadPoll function. The existing code failed due to not properly handling the very finicky buffer sizes related to the LobPrefetch size (which defaults to 8k in Zeos.) However I've replaced that call with a non-polling version named ReadFull which is simpler and safer. I've left both versions in just in case there's a need.


........